perf(ui): serve trees from one scoped scan; share the Eloquent scope concern
EloquentTreeSource issued two queries per level under the eager Tree
walk (~2N per render on deep hierarchies); it now loads the scoped
table once into an adjacency map and answers roots/children/hasChildren
from it. The scope/query plumbing it duplicated from EloquentOptions
lives in a shared Core\Concerns\ScopedEloquentQuery, ready for the next
Eloquent-backed source.
