+++ src/Core/EloquentOptions.php
@@ -7,6 +7,7 @@ use Closure;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
+use Lattice\Lattice\Core\Concerns\ScopedEloquentQuery;
 use Lattice\Lattice\Core\Contracts\OptionSource;
 
 /**
@@ -16,11 +17,10 @@ use Lattice\Lattice\Core\Contracts\OptionSource;
  */
 final class EloquentOptions implements OptionSource
 {
+    use ScopedEloquentQuery;
+
     /** @var list<string>|null */
     private ?array $searchColumns = null;
-
-    /** @var Closure(Builder<Model>): mixed|null */
-    private ?Closure $scope = null;
 
     private int $limit = 50;
 
@@ -73,18 +73,6 @@ final class EloquentOptions implements OptionSource
         return $this;
     }
 
-    /**
-     * Constrain the query (e.g. only active rows). Applied to both search and selected.
-     *
-     * @param  Closure(Builder<Model>): mixed  $scope
-     */
-    public function scope(Closure $scope): self
-    {
-        $this->scope = $scope;
-
-        return $this;
-    }
-
     public function limit(int $limit): self
     {
         $this->limit = $limit;
@@ -133,24 +121,6 @@ final class EloquentOptions implements OptionSource
     private function valueColumn(): string
     {
         return $this->resolvedValueKey ??= $this->valueKey ?? (new $this->model)->getKeyName();
-    }
-
-    /**
-     * @return Builder<Model>
-     */
-    private function query(): Builder
-    {
-        $builder = $this->model::query();
-
-        if ($this->scope instanceof Closure) {
-            $scoped = ($this->scope)($builder);
-
-            if ($scoped instanceof Builder) {
-                $builder = $scoped;
-            }
-        }
-
-        return $builder;
     }
 
     /**
